+++ Abc.MoqComplete/Abc.MoqComplete/CompletionProvider/AutoMockerCallbackMethodProvider.cs
@@ -1,77 +1,29 @@
-﻿using System.Linq;
-using Abc.MoqComplete.Extensions;
+﻿using System.Collections.Generic;
 using Abc.MoqComplete.Services;
 using JetBrains.ProjectModel;
-using JetBrains.ReSharper.Feature.Services.CodeCompletion;
-using JetBrains.ReSharper.Feature.Services.CodeCompletion.Infrastructure;
-using JetBrains.ReSharper.Feature.Services.CodeCompletion.Infrastructure.LookupItems;
-using JetBrains.ReSharper.Feature.Services.CSharp.CodeCompletion.Infrastructure;
-using JetBrains.ReSharper.Features.Intellisense.CodeCompletion.CSharp;
 using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.CSharp;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
-using JetBrains.ReSharper.Psi.ExpectedTypes;
-using JetBrains.ReSharper.Psi.Resources;
-using JetBrains.ReSharper.Psi.Tree;
 
 namespace Abc.MoqComplete.CompletionProvider
 {
 	[Language(typeof(CSharpLanguage))]
-	public class AutoMockerCallbackMethodProvider : ItemsProviderOfSpecificContext<CSharpCodeCompletionContext>
+	public class AutoMockerCallbackMethodProvider : BaseCallbackMethodProvider
 	{
-		protected override bool IsAvailable(CSharpCodeCompletionContext context)
+		/// <inheritdoc />
+		protected override IMethod GetMockedMethodFromSetupMethod(ISolution solution, IInvocationExpression invocation)
 		{
-			var codeCompletionType = context.BasicContext.CodeCompletionType;
+			var mockedMethodProvider = solution.GetComponent<IAutoMockerMockedMethodProvider>();
 
-			return codeCompletionType == CodeCompletionType.SmartCompletion || codeCompletionType == CodeCompletionType.BasicCompletion;
+			return mockedMethodProvider.GetMockedMethodFromSetupMethod(invocation);
 		}
 
-		protected override bool AddLookupItems(CSharpCodeCompletionContext context, IItemsCollector collector)
+		/// <inheritdoc />
+		protected override IEnumerable<string> GetMockedMethodParameterTypes(ISolution solution, IInvocationExpression invocation)
 		{
-			var identifier = context.TerminatedContext.TreeNode as IIdentifier;
-			var expression = identifier.GetParentSafe<IReferenceExpression>();
+			var mockedMethodProvider = solution.GetComponent<IAutoMockerMockedMethodProvider>();
 
-			if (expression == null)
-			{
-				return false;
-			}
-
-			if (!(expression.ConditionalQualifier is IInvocationExpression invocation))
-			{
-				return false;
-			}
-
-			var methodIdentifier = context.BasicContext.Solution.GetComponent<IMoqMethodIdentifier>();
-			var mockedMethodProvider = context.BasicContext.Solution.GetComponent<IAutoMockerMockedMethodProvider>();
-
-			if (methodIdentifier.IsMoqReturnMethod(invocation))
-			{
-				invocation = invocation.InvokedExpression?.FirstChild as IInvocationExpression;
-			}
-
-			var mockedMethod = mockedMethodProvider.GetMockedMethodFromSetupMethod(invocation);
-
-			if (mockedMethod == null || mockedMethod.Parameters.Count == 0)
-			{
-				return false;
-			}
-
-			var types = mockedMethodProvider.GetMockedMethodParameterTypes(invocation);
-			var variablesName = mockedMethod.Parameters.Select(p => p.ShortName);
-			var proposedCallback = $"Callback<{string.Join(",", types)}>(({string.Join(",", variablesName)}) => {{}})";
-
-			var item = CSharpLookupItemFactory.Instance.CreateKeywordLookupItem(context,
-				proposedCallback,
-				TailType.None,
-				PsiSymbolsThemedIcons.Method.Id);
-
-			item.SetInsertCaretOffset(-2);
-			item.SetReplaceCaretOffset(-2);
-			item.WithInitializedRanges(context.CompletionRanges, context.BasicContext);
-			item.SetTopPriority();
-			collector.Add(item);
-
-			return true;
+			return mockedMethodProvider.GetMockedMethodParameterTypes(invocation);
 		}
 	}
 }
